+++ base/Enums.jl
@@ -143,7 +143,7 @@ macro enum(T, syms...)
                 print(io, x)
             else
                 print(io, x, "::")
-                showcompact(io, typeof(x))
+                show(IOContext(io, :compact => true), typeof(x))
                 print(io, " = ", $basetype(x))
             end
         end