+++ base/deprecated.jl
@@ -1400,6 +1400,10 @@ end
     Base.@deprecate log Base.log
 end
 
+@deprecate showcompact(x) show(IOContext(STDOUT, :compact => true), x)
+@deprecate showcompact(io, x) show(IOContext(io, :compact => true), x)
+@deprecate sprint(::typeof(showcompact), args...) sprint(show, args...; context=:compact => true)
+
 # END 0.7 deprecations
 
 # BEGIN 1.0 deprecations