+++ base/exports.jl
@@ -624,7 +624,6 @@ export
     println,
     printstyled,
     show,
-    showcompact,
     showerror,
     sprint,
     summary,